+++ components/sign/types.ts
@@ -1,0 +1,6 @@
+export interface PostSignProps {
+  email: string;
+  fullName: string;
+  password: string;
+  onSuccess: () => void;
+}
