+++ pages/profile/index.tsx
@@ -1,24 +1,9 @@
-import { getLocalstorage } from '@components/auth/localstorage';
-import { getId } from '@components/profile/getId';
-import Profile from '@components/profile/index';
-import { publicApi } from 'api';
+import { getUserData } from '@components/profile/api/getUserData';
 import { useQuery } from 'react-query';
+import Profile from '@components/profile/index';
 
 export default function MyProfile() {
-  const getData = async () => {
-    try {
-      const token = getLocalstorage('JWT_TOKEN');
-      const id = await getId(token);
-      const response = await publicApi.get(`/users/${id}`);
-      if (response.status === 200) {
-        return response.data;
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  const { status, data } = useQuery('myData', getData);
+  const { status, data } = useQuery('myData', getUserData);
 
   if (status === 'loading') {
     return <span>Loading...</span>;
